+++ packages/formatters/esbuild.ts
@@ -1,8 +1,13 @@
 import { build } from 'esbuild'
 import { readdirSync } from 'fs'
-import { resolve } from 'path'
+import { dirname, resolve } from 'path'
+import { fileURLToPath } from 'url'
 
 const watch = process.argv.includes('--watch')
+
+//@ts-ignore
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = dirname(__filename)
 
 const getPath = (file: string) => resolve(__dirname, file)
 
@@ -15,7 +20,7 @@ files.forEach((file) =>
 		build({
 			entryPoints: [`./src/${file}`],
 			bundle: true,
-			outfile: getPath(`../../formatters/${file.replace('.ts', `.${format === 'esm' ? 'm' : 'c'}js`)}`),
+			outfile: getPath(`../../formatters/${file.replace('.mts', `.${format === 'esm' ? 'm' : 'c'}js`)}`),
 			platform: 'neutral',
 			format,
 			minify: true,
