+++ packages/formatters/generateTreeShakableIndexFile.ts
@@ -1,12 +1,17 @@
 /* eslint-disable no-console */
 
 import { readFileSync, writeFileSync } from 'fs'
-import { resolve } from 'path'
+import { dirname, resolve } from 'path'
+import { fileURLToPath } from 'url'
+
+//@ts-ignore
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = dirname(__filename)
 
 console.log(`
 creating tree-shakable index files ...`)
 
-const content = readFileSync(resolve(__dirname, './src/index.ts'))
+const content = readFileSync(resolve(__dirname, './src/index.mts'))
 	.toString()
 	.split(/\r?\n/)
 	.filter((line) => !line.startsWith('export type'))
