+++ packages/formatters/package.json
@@ -13,5 +13,6 @@
     "typescript": "^4.7.4",
     "uvu": "^0.5.6",
     "watchlist": "^0.3.1"
-  }
+  },
+  "type": "module"
 }
