+++ packages/formatters/src/identity.mts
@@ -1,0 +1,2 @@
+// eslint-disable-next-line prettier/prettier
+export default <T,>(value: T): T => value
