--- packages/formatters/src/identity.ts
@@ -1,1 +1,0 @@
-export default <T>(value: T): T => value
