--- packages/formatters/src/index.ts
@@ -1,8 +1,0 @@
-export { default as date } from './date'
-export { default as identity } from './identity'
-export { default as ignore } from './ignore'
-export { default as lowercase } from './lowercase'
-export { default as number } from './number'
-export { default as replace } from './replace'
-export { default as time } from './time'
-export { default as uppercase } from './uppercase'
