+++ packages/formatters/src/lowercase.mts
