+++ packages/formatters/src/number.mts
