+++ packages/formatters/src/uppercase.mts
