+++ packages/formatters/test/formatters.test.ts
@@ -1,7 +1,7 @@
 import { i18nObject, i18nString } from '@typesafe-i18n/runtime/index.mjs'
 import { suite } from 'uvu'
 import * as assert from 'uvu/assert'
-import { date, identity, ignore, lowercase, number, replace, uppercase } from '../src'
+import { date, identity, ignore, lowercase, number, replace, uppercase } from '../src/index.mjs'
 
 const test = suite('formatters')
